Add NAT GW to public subnets, change IGW name
Add 2 NAT Gateways in different AZs for redundancy in the public subnet
so that they can reach the public internet. These will be used for egress
from private networks.

As well, fix the IGW name tag to better match convention.

This also moves the data resource that finds availability zones in AWS
for assignment to subnets to the vpc config, and out of the public
network config.